requirements.txt
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
tensorflow
pillow

# tests/linter
# black ?
# coverage ?
# flake8 ?
# pytest ?
# yapf ?

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
# time


# streamlit
streamlit
opencv-python
python-dotenv
